SyncFix/requirements.txt
pymupdf>=1.23.0
sentence-transformers>=2.7.0
chromadb>=0.5.0
transformers>=4.40.0
torch==2.6.0
streamlit>=1.35.0
Pillow>=10.0.0
rank-bm25>=0.2.2
requests>=2.31.0

numpy>=1.24.0
tqdm>=4.65.0